添加mybatis-plus 数据权限

+++ src/main/java/com/wl/security_demo/domain/entity/SysRole.java
@@ -61,5 +61,5 @@ public class SysRole implements Serializable {
      * 1:全部, 2:自定义, 3:本部门, 4:本部门及以下, 5:仅本人
      */
     @TableField("data_scope")
-    private Byte dataScope;
+    private Integer dataScope;
 }
